+++ engine/flink/management/periodic/src/main/scala/pl/touk/nussknacker/engine/management/periodic/PeriodicProcessException.scala
@@ -1,3 +1,5 @@
 package pl.touk.nussknacker.engine.management.periodic
 
-class PeriodicProcessException(message: String) extends RuntimeException(message)
+class PeriodicProcessException(message: String, parent: Throwable) extends RuntimeException(message, parent) {
+  def this(message: String) = this(message, null)
+}